refactor: Created repository interface to prepare for gitlab addition
